refactor: split sqlite download task store
